fraud_detection_module/fraud_detection_system/fastapi_app/requirements.txt
fastapi
uvicorn
joblib
scikit-learn
xgboost
shap
pandas
kafka-python 